prediction-api/requirements.txt
flask==2.3.0
flask-jwt-extended==4.5.3
flask-cors==4.0.0
numpy==1.24.0
pandas==2.0.0
scikit-learn==1.5.1
joblib==1.3.0
google-generativeai==0.3.0
python-dotenv==1.0.0
werkzeug==2.3.7
requests==2.31.0
authlib==1.2.0
